helpdesk/requirements.txt
Django
djangorestframework
django-cors-headers
Pillow
djangorestframework-simplejwt
celery
google-api-python-client>=2.85.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0